KUKA-0009_Implement_Customer_Dao

See merge request Alex_Strutinszky/h2gKukaTeamproject!10

+++ customermanagementsvc/modules/domain/src/main/java/com/kuka/h2g/domain/customer/Customer.java
@@ -8,7 +8,7 @@ import lombok.Value;
 @Builder(setterPrefix = "with")
 public class Customer {
 
-    int customerId;
+    long customerId;
     String firstName;
     String secondName;
     Address address;
